pyproject.toml
[project]
name = "clarus"
version = "0.1.0"
description = "A Flask-based RAG (Retrieval Augmented Generation) application"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Alif Ramadhan", email = "alif.datascientist@gmail.com" }
]
dependencies = [
    "flask>=3.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-groq>=0.1.0",
    "chromadb>=0.4.0",
    "pypdf2>=3.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/NnA301023/clarus"
Repository = "https://github.com/NnA301023/clarus.git"
